ft-163046711 modified the model structure
